brocolib::runtime_metadata

Struct Il2CppMetadataRegistration

pub struct Il2CppMetadataRegistration {
    pub generic_classes: Vec<Il2CppGenericClass>,
    pub generic_insts: Vec<Il2CppGenericInst>,
    pub generic_method_table: Vec<Il2CppGenericMethodFunctionsDefinitions>,
    pub types: Vec<Il2CppType>,
    pub array_types: Vec<Il2CppArrayType>,
    pub method_specs: Vec<Il2CppMethodSpec>,
    pub field_offsets: Option<Vec<Vec<u32>>>,
    pub type_definition_sizes: Option<Vec<Il2CppTypeDefinitionSizes>>,
}

Defined at il2cpp-class-internals.h:622

Fields
generic_classes: Vec<Il2CppGenericClass>generic_insts: Vec<Il2CppGenericInst>generic_method_table: Vec<Il2CppGenericMethodFunctionsDefinitions>types: Vec<Il2CppType>array_types: Vec<Il2CppArrayType>

This is not a real field in the metadata. It is here to provide the ability to access array types by index instead of by pointer.

method_specs: Vec<Il2CppMethodSpec>field_offsets: Option<Vec<Vec<u32>>>

Compiler calculated field offset values. Only exists when read from an ELF. Since this is platform dependent, it cannot be read from C++ sources.

type_definition_sizes: Option<Vec<Il2CppTypeDefinitionSizes>>

Compiler calculated size values. Only exists when read from an ELF. Since this is platform dependent, it cannot be read from C++ sources.

Implementations

impl Il2CppMetadataRegistration

pub fn read( reader: &ObjectReader<'_>, addr: u64, metadata: &GlobalMetadata<'_>, ) -> Result<Self>

impl Il2CppMetadataRegistration

pub fn read_src( src_dir: &SourceDir, name_mappings: &NameMappings<'_>, ) -> Result<Self, SourceParseError>