brocolib::runtime_metadata

Struct Il2CppMethodSpec

pub struct Il2CppMethodSpec {
    pub method_definition_index: MethodIndex,
    pub class_inst_index: u32,
    pub method_inst_index: u32,
}

A generic method instantiation.

It is not possible for both class_inst_index and method_inst_index to be invalid since if both the class and method are not generic, you cannot make a generic instance.

Defined at il2cpp-metadata.h:67

Fields
method_definition_index: MethodIndex

The method definition.

class_inst_index: u32

The class generic argument list (if class is generic).

Indices into the Il2CppMetadataRegistration::generic_insts field

method_inst_index: u32

The method generic argument list (if method is generic).

Indices into the Il2CppMetadataRegistration::generic_insts field