brocolib::runtime_metadata

Module elf

ELF runtime metadata parsing.

For IL2CPP Unity games that are built for linux or linux-based platforms, you will find a shared object file named libil2cpp.so. This file contains the libil2cpp library, code generated from C#, as well as certain metadata information.

To read metadata information from libil2cpp.so, see RuntimeMetadata::read() and RuntimeMetadata::read_elf().

Structs

DisassembleError

Enums

Il2CppBinaryError

Functions

addr_in_bss

get_str

strlen

vaddr_conv

Converts a virtual address in the elf to a file offset

Type Aliases

Elf